requirements.txt
langchain
validators
python-dotenv
langchain-groq
beautifulsoup4
python-magic
youtube-transcript-api
pytube
langchain-community
fastapi